Move the YAML front matter check into a separate file
