+++ scripts/linter.rb
@@ -12,6 +12,7 @@ require_relative 'linting/logging'
 
 require_relative 'linting/check_all_urls_are_hackable'
 require_relative 'linting/check_with_html_proofer'
+require_relative 'linting/check_yaml_front_matter'
 
 # This checks that every article on /elsewhere/ has at least one copy
 # archived on my own computers.
@@ -50,51 +51,6 @@ def get_markdown_paths(src_dir)
     .reject { |md_path| md_path == "#{src_dir}/400.md" }
 end
 
-# Validate the YAML front matter by checking that:
-#
-#   1. I'm not using undocumented fields
-#   2. Fields have appropriate values
-#
-def check_yaml_front_matter(src_dir)
-  errors = Hash.new { [] }
-
-  info('Checking YAML front matter...')
-
-  schema = JSON.parse(File.read('front-matter.json'))
-
-  get_markdown_paths(src_dir).each do |md_path|
-    # The YAML loader will try to be "smart" (e.g. reading dates as
-    # proper Ruby date types), which is unhelpful for json-schema checking.
-    #
-    # Make sure everything is JSON-esque (i.e. strings/numbers/bools)
-    # before passing to the json-schema gem.
-    front_matter = YAML.load(
-      File.read(md_path).split("\n---\n")[0],
-      permitted_classes: [Date, Time]
-    )
-    front_matter = JSON.parse(JSON.dump(front_matter))
-
-    md_errors = JSON::Validator.fully_validate(schema, front_matter)
-
-    errors[md_path] = md_errors unless md_errors.empty?
-
-    expected_layout =
-      if md_path.start_with?("#{src_dir}/_posts") || md_path.start_with?("#{src_dir}/_drafts")
-        'post'
-      elsif md_path.start_with?('src/_til')
-        'til'
-      else
-        'page'
-      end
-
-    if front_matter['layout'] != expected_layout
-      errors[md_path] <<= "layout should be '#{expected_layout}'; got #{front_matter['layout']}"
-    end
-  end
-
-  report_errors(errors)
-end
-
 # Check my Netlify redirects point to real pages.
 #
 # This ensures that any redirects I create are working.  It doesn't mean
@@ -130,21 +86,6 @@ def check_redirects(dst_dir)
   bad_lines.each do |ln|
     lineno, line = ln
     error("  * L#{lineno}:\t#{line}")
-  end
-  exit!
-end
-
-def report_errors(errors)
-  # This is meant to look similar to the output from HTMLProofer --
-  # errors are grouped by filename, so they can be easily traced
-  # back to the problem file.
-  return if errors.empty?
-
-  errors.each do |display_path, messages|
-    error("- #{display_path}")
-    messages.each do |m|
-      error("  *  #{m}")
-    end
   end
   exit!
 end
